requirements.txt
Cython==0.25.2
appdirs==1.4.3
Kivy==1.9.1
Kivy-Garden==0.1.4
packaging==16.8
plyer==1.3.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
websocket-client-py3==0.15.0
